pyproject.toml
[build-system]
requires = ["hatchling>=1.21"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/pydantic_ai_gepa"]

[tool.hatch.build.targets.sdist]
include = ["src/pydantic_ai_gepa", "README.md", "pyproject.toml"]

[tool.hatch.metadata]
allow-direct-references = true

[project]
name = "pydantic-ai-gepa"
version = "0.1.0"
description = "GEPA optimization integration for pydantic-ai"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "cloudpickle>=3.0.0",
    "pydantic-ai-slim[openai,anthropic,logfire,retries]>=1.96.0",
    "pydantic-graph>=1.96.0",
    "pydantic-evals>=1.96.0",
    "rich>=13.7.1",
    "logfire>=4.14.2",
    "PyYAML>=6.0.0",
    "pydantic-monty>=0.0.17",
    "typer>=0.12.0",
]

[project.scripts]
gepa = "pydantic_ai_gepa.cli:app"

[dependency-groups]
dev = [
    "pytest>=8.0.0",
    "pytest-cov>=5.0.0",
    "inline-snapshot>=0.29.0",
    "pytest-asyncio>=1.2.0",
    "ipdb>=0.13.13",
    "logfire>=4.14.2",
    "pyright>=1.1.405",
    "time-machine>=2.19.0",
    "ruff>=0.6.9",
    "pydantic-ai-slim[temporal]>=1.96.0",
]

[tool.inline-snapshot]
# Inline-snapshot expects a format string; provide ruff invocation with filename placeholder.
format-command = "ruff format --stdin-filename {filename} -"
